+++ src/browser/__tests__/test-theme.js
@@ -1,0 +1,8 @@
+describe('theme', () => {
+    jest.unmock('../theme');
+    const theme = require('../theme').default;
+
+    it('is object', () => {
+        expect(typeof(theme)).toEqual('object');
+    });
+});