+++ src/browser/theme.js
@@ -1,0 +1,10 @@
+import Colors from "material-ui/lib/styles/colors";
+
+export default {
+    fontFamily: '"Yu Gothic", "Meiryo", "Noto Sans Japanese", sans-serif',
+    palette: {
+        primary1Color: Colors.green500,
+        primary2Color: Colors.green700,
+        accent1Color: Colors.purpleA200,
+    },
+};